vibes is now calculated based on album/artist genres bc of deprecacted api endpoint
