+++ frontend/src/components/data/SpotifyAuth.tsx
@@ -5,7 +5,7 @@ import axios from 'axios';
 export const SpotifyLoginButton: FunctionComponent = () => {
   const CLIENT_ID = process.env.REACT_APP_CLIENT_ID;
   const AUTH_URL = `https://accounts.spotify.com/authorize?client_id=${CLIENT_ID}&response_type=code&redirect_uri=http://localhost:3000/profile&scope=user-read-email%20user-read-private%20user-library-read%20user-library-modify%20user-follow-read%20user-follow-modify`;
-
+  
   return (
     <Button
       sx={{
