+++ package.json
@@ -21,7 +21,9 @@
     "@mui/icons-material": "^6.1.5",
     "@mui/material": "^6.4.4",
     "axios": "^1.7.7",
-    "prettier": "^3.3.3"
+    "prettier": "^3.3.3",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.14.0",
